Use type_for to identify the type of cached attribute
